Added: SettingsAndConstants.java Description: Holds constants that may be used elsewhere multiple times. Or just large strings.
Added: ServerClientThread.java
Description: Acts as the clients connection to the CentralServer. This is required for multiple connections to the CentralServer.

Renamed: ProfileServer.class to CentralServer.java
Description: As this is the central server used to run the whole platform, I feel like CentralServer would be a more understandable.

Changed: UserClient.java
Description: Recreated the simple client script but using the new ServerClientThread.java system.

Added: ServerAndClient package
Description: This is just so all of the client and server things are organized. We can do this to more of the files as we go on and if you need to use something from it, all you have to do is import the package at the top. (Intellij will do this/tell you automatically)